% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatastRoEus-package.R
\docType{package}
\name{CatastRoEus-package}
\alias{CatastRoEus}
\alias{CatastRoEus-package}
\title{CatastRoEus: Interface to the 'INSPIRE' services of 'Catastro de Bizkaia', 'Catastro de Gipuzkua', 'Catastro de Araba'}
\description{
Access public spatial data available under the 'INSPIRE' directive. Tools for downloading references, buildings and addresses of properties on Pais Vasco (Spain).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rOpenSpain/CatastRoEus}
  \item Report bugs at \url{https://github.com/rOpenSpain/CatastRoEus/issues}
}

}
\author{
\strong{Maintainer}: Jimena Salinero \email{jimenasalivega@gmail.com}

}
\keyword{internal}
