% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_eus_bu.R
\name{catreus_wfs_get_buildings_bbox}
\alias{catreus_wfs_get_buildings_bbox}
\title{Retrieve Building Data in Pais Vasco Based on Bounding Box Coordinates}
\usage{
catreus_wfs_get_buildings_bbox(x, srs, verbose = FALSE, count = 10)
}
\arguments{
\item{x}{Bounding box coordinates or a spatial object, which could be:
\itemize{
\item A numeric vector of length 4 with the coordinates defining the bounding box:
\code{c(latitude1, longitude1, latitude2, longitude2)}.
\item An \code{sf/sfc} object from the \CRANpkg{sf} package.
}}

\item{srs}{Spatial Reference System (SRS) or Coordinate Reference System (CRS) code
to be used in the query. For best results, ensure the coordinates are in ETRS89
(EPSG:4258) or WGS84 (EPSG:4326) when using latitude and longitude.}

\item{verbose}{Logical; if \code{TRUE}, additional information about function operations
is printed. Useful for debugging. Default is \code{FALSE}.}

\item{count}{Integer specifying the maximum number of building records to return.
Default is 10.}
}
\value{
Depending on the input and geographic location, this function may return:
\itemize{
\item An \code{sf} object containing building data within the specified bbox.
\item A message indicating mismatched or out-of-region coordinates.
}
}
\description{
Fetches the spatial data of buildings within a specified bounding box. The function
determines the province based on the provided coordinates and fetches building data
for Bizkaia, Gipuzkoa, or Araba/Álava accordingly. It checks if the bounding box
falls within a single province and handles different CRS inputs.
}
\details{
This function uses reverse geocoding to determine the province within the Basque Country
from the coordinates provided. Based on the province, it delegates the data fetching to
specific functions handling each province's data. It supports flexible input types and
handles geographical coordinate transformations internally if needed.
}
\examples{
\donttest{
# Define bounding box coordinates for a location in Gipuzkoa
coords_gipuzkoa <- c(42.994337, -2.554863, 43.344138, -1.779831)

# Fetch building data using the bounding box
buildings_gipuzkoa <- catreus_wfs_get_buildings_bbox(coords_gipuzkoa, srs = 25830, count = 10)

library(ggplot2)
# Plot the buildings data
ggplot(buildings_gipuzkoa) +
  geom_sf() + ggtitle("Building Data for Gipuzkoa")

# Define bounding box coordinates for a location in Bizkaia
coords_bizkaia <- c(43.312, -2.994, 43.313, -2.993)

# Fetch building data using the bounding box
buildings_bizkaia <- catreus_wfs_get_buildings_bbox(coords_bizkaia, srs = 4326, count = 10)

library(ggplot2)
# Plot the buildings data for Bizkaia
ggplot(buildings_bizkaia) +
  geom_sf() + ggtitle("Building Data for Bizkaia")

# Define bounding box coordinates for a rural location in Araba/Álava
coords_araba <- c(539226.596, 4744012.338, 539236.286, 4744133.782)

# Fetch building data using the bounding box, requesting more features
buildings_araba <- catreus_wfs_get_buildings_bbox(coords_araba, srs = 25830, count = 20)

library(ggplot2)
# Plot the buildings data for Araba/Álava
ggplot(buildings_araba) +
  geom_sf() + ggtitle("Building Data for Araba/Álava")

}
  
}
\seealso{
\code{\link[sf:st_bbox]{st_bbox}}, which is used to manage spatial bounding boxes.
}
